Refactor envelope_with_default_community to use MountableAPI
