+++ app/api/v1/base.rb
@@ -5,10 +5,7 @@ require 'v1/root'
 require 'v1/schemas'
 require 'v1/search'
 require 'v1/ce_registry'
-require 'v1/resource_api'
 require 'v1/resources'
-require 'v1/community_resources'
-require 'v1/community_envelopes'
 require 'v1/envelopes'
 require 'v1/publishers'
 require 'v1/organizations'
@@ -31,14 +28,13 @@ module API
       mount API::V1::Search
       mount API::V1::CERegistry
       mount API::V1::Resources
+      mount API::V1::Resources.api_class
+      mount API::V1::Envelopes.api_class
       mount API::V1::Envelopes
 
-      route_param :envelope_community do
-        mount API::V1::CommunityEnvelopes
-      end
-
       route_param :community_name do
-        mount API::V1::CommunityResources
+        mount API::V1::Resources.api_class
+        mount API::V1::Envelopes.api_class
       end
 
       namespace :metadata do
